requirements.txt
Django~=5.2.4
structlog~=25.4.0
python-decouple~=3.8
djangorestframework~=3.16.0
openai~=1.97.1
gunicorn~=21.2.0
